+++ NEWS.md
@@ -1,5 +1,7 @@
 # medfate 5.1.1
 * Update of trait_family_means (avoiding infinite mean values)
+* Preformed leaf area in growth
+* Minimum and maximum values added to SpParamsDefinition
 * Bug correction:
     - RespDist/RespFire initialization for shrubs
     - Resprouting for shrubs